Fork `SlotId` into `SlotId` and `EpochSlotId`.
A `SlotId` is a simple 0-based linear count of the number of slots since
the genesis slot (which we assign a value of 0). We use values of this
type within the wallet and database layers, for simplicity.

An `EpochSlotId` is a compound value of (epoch number, slot number).
We use values of this type in API functions, and when communicating with
various node backends.

Values can be converted from one type to another with the following
functions:

  1. slotIdToEpochSlotId
  2. epochSlotIdToSlotId